fixed(display): Using StrokeDashArray with very long lines could result in artifacts.

+++ packages/display/src/displays/webgl/buffer/templates/LineBuffer.ts
@@ -43,7 +43,7 @@ export class LineBuffer extends TemplateBuffer {
                 size: 4
             },
             a_lengthSoFar: {
-                data: new FlexArray(Uint16Array),
+                data: new FlexArray(Float32Array),
                 size: 1
             }
         };
